requirements.txt
biopython==1.79
matplotlib==3.7.0
numpy==1.23.5
pandas==1.5.3
pydantic
python-dotenv
PyYAML
tqdm
wandb